afroyacaCore/requirements.txt
backcall==0.2.0
decorator==4.4.2
Django==2.2.17
django-rest-framework==0.1.0
djangorestframework==3.12.2
ipython==7.9.0
ipython-genutils==0.2.0
jedi==0.17.2
parso==0.7.1
pexpect==4.8.0
pickleshare==0.7.5
pkg-resources==0.0.0
prompt-toolkit==2.0.10
psycopg2==2.8.6
ptyprocess==0.6.0
Pygments==2.7.3
pytz==2020.4
six==1.15.0
sqlparse==0.4.1
traitlets==4.3.3
wcwidth==0.2.5
backcall==0.2.0
decorator==4.4.2
Django==2.2.17
django-cors-headers==3.5.0
django-rest-framework==0.1.0
djangorestframework==3.12.2
ipython==7.9.0
ipython-genutils==0.2.0
jedi==0.17.2
parso==0.7.1
pexpect==4.8.0
pickleshare==0.7.5
pkg-resources==0.0.0
prompt-toolkit==2.0.10
psycopg2==2.8.6
ptyprocess==0.6.0
Pygments==2.7.3
pytz==2020.4
six==1.15.0
sqlparse==0.4.1
traitlets==4.3.3
wcwidth==0.2.5
